Move IgnoreOnceQuickFixTests for QuickFixes to RD.QuickFixes
