+++ RubberduckTests/QuickFixes/IgnoreOnceQuickFixTests.cs
@@ -191,5 +191,38 @@ End Sub";
             new IgnoreOnceQuickFix(state, new[] { inspection }).Fix(inspectionResults.First());
             Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
         }
+
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void AnnotationListFollowedByCommentAddsAnnotationCorrectly()
+        {
+            const string inputCode = @"
+Public Function GetSomething() As Long
+    '@Ignore VariableNotAssigned: Is followed by a comment.
+    Dim foo
+    GetSomething = foo
+End Function
+";
+
+            const string expectedCode = @"
+Public Function GetSomething() As Long
+    '@Ignore VariableTypeNotDeclared, VariableNotAssigned: Is followed by a comment.
+    Dim foo
+    GetSomething = foo
+End Function
+";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new VariableTypeNotDeclaredInspection(state);
+            var inspectionResults = inspection.GetInspectionResults();
+
+            new IgnoreOnceQuickFix(state, new[] { inspection }).Fix(inspectionResults.First());
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
+
     }
 }
